sorry huge commit! added base for API, split many functions to be more modular (for API), added support for r5 v4 with INT's as the osmids type, also changed the segmentation OSM ID creation logic to produce smaller numbers that r5 can use INT's instead of Long's

+++ green_paths_2/src/database_controller.py
@@ -1,7 +1,7 @@
 import os
 from collections import defaultdict
 import sqlite3
-from typing import Dict, List, Any
+from typing import Dict, List, Any, Optional, Tuple
 
 from ..src.config import GP2_DB_PATH, GP2_DB_TEST_PATH, SEGMENT_STORE_TABLE
 from ..src.timer import time_logger
@@ -20,6 +20,14 @@ class DatabaseController:
     def connect(self):
         return sqlite3.connect(self.db_path)
 
+    # empty table
+    def empty_table(self, table: str):
+        conn = self.connect()
+        cursor = conn.cursor()
+        cursor.execute(f"DELETE FROM {table}")
+        conn.commit()
+        conn.close()
+
     def drop_table(self, table: str):
         conn = self.connect()
         cursor = conn.cursor()
@@ -27,11 +35,15 @@ class DatabaseController:
         conn.commit()
         conn.close()
 
-    def create_table_from_dict_data(self, table: str, data: Dict[str, Any]):
-        """Creates db tables based on dict data"""
+    # TODO: put force as parameter bool
+    def create_table_from_dict_data(
+        self, table: str, data: Dict[str, Any], force: bool = False
+    ):
+        """Drops targe table if exists and then creates db tables based on dict data"""
         # drop the table if it already exists
         # for the tables can have different columns
-        self.drop_table(table)
+        if force:
+            self.drop_table(table)
         columns = []
         for key, value in data.items():
             if key == "osm_id":
@@ -111,7 +123,8 @@ class DatabaseController:
         for i in range(0, len(values_list), chunk_size):
             chunk = values_list[i : i + chunk_size]
             cursor.executemany(
-                f"INSERT INTO {table} ({columns}) VALUES ({placeholders})", chunk
+                f"INSERT OR IGNORE INTO {table} ({columns}) VALUES ({placeholders})",
+                chunk,
             )
             conn.commit()
         conn.close()
@@ -223,11 +236,24 @@ class DatabaseController:
     # get multiple records by column values
 
     # Function to select rows based on one or many osm_id values
-    def select_rows_by_osm_ids(self, table: str, osm_ids: List[int]) -> List[dict]:
+    # Function to select rows based on one or many osm_id values
+    def select_rows_by_osm_ids(
+        self,
+        table: str,
+        osm_ids: List[int],
+        columns: Optional[List[str]] = None,  # Optional parameter for column selection
+    ) -> List[dict]:
         conn = self.connect()
         cursor = conn.cursor()
+
+        # Determine which columns to select
+        if columns:
+            columns_str = ", ".join(columns)
+        else:
+            columns_str = "*"
+
         placeholders = ", ".join(["?"] * len(osm_ids))
-        query = f"SELECT * FROM {table} WHERE osm_id IN ({placeholders})"
+        query = f"SELECT {columns_str} FROM {table} WHERE osm_id IN ({placeholders})"
         cursor.execute(query, osm_ids)
         rows = cursor.fetchall()
         conn.close()
@@ -244,15 +270,24 @@ class DatabaseController:
         conn.close()
         return [dict(zip(columns, row)) for row in rows]
 
-    def get_all(self, table: str, column_names: bool = False) -> List[Dict[str, Any]]:
+    def get_all(
+        self, table: str, column_names: bool = False
+    ) -> Tuple[List[Dict[str, Any]], List[str]]:
         conn = self.connect()
         cursor = conn.cursor()
         cursor.execute(f"SELECT * FROM {table}")
         rows = cursor.fetchall()
+
+        # Fetch column names if required
         if column_names:
             column_names = [description[0] for description in cursor.description]
+            # Convert rows to list of dictionaries
+            rows = [dict(zip(column_names, row)) for row in rows]
         else:
             column_names = None
+            # Otherwise keep rows as list of tuples
+            rows = [tuple(row) for row in rows]
+
         conn.close()
         return rows, column_names
 
@@ -264,37 +299,81 @@ class DatabaseController:
         count = cursor.fetchone()[0]
         return count
 
-    # create index for table
-    def create_index(self, table: str, column: str) -> None:
+    def create_index(self, table, column):
+        # Query to check if the index already exists
         conn = self.connect()
         cursor = conn.cursor()
-        cursor.execute(f"CREATE INDEX {column} ON {table}({column});")
-        conn.commit()
-        conn.close()
-
-    def add_column_to_table(self, table_name, column_name, column_type="TEXT"):
-        """
-        Add a new column to an existing table in the database.
-
-        Parameters:
-        table_name (str): The name of the table to alter.
-        column_name (str): The name of the new column to add.
-        column_type (str): The type of the new column (default is TEXT).
-        """
-        conn = self.connect()
-        cursor = conn.cursor()
-
-        alter_table_sql = (
-            f"ALTER TABLE {table_name} ADD COLUMN {column_name} {column_type};"
+        index_name = (
+            f"{column}_index"  # You might have a naming convention for your indexes
         )
+        # Check if the index already exists
+        cursor.execute(f"PRAGMA index_list({table});")
+        existing_indexes = [
+            index[1] for index in cursor.fetchall()
+        ]  # index[1] is the index name in the result
 
-        try:
-            cursor.execute(alter_table_sql)
+        if index_name not in existing_indexes:
+            # Only create the index if it doesn't already exist
+            cursor.execute(f"CREATE INDEX {index_name} ON {table}({column});")
             conn.commit()
-        except sqlite3.OperationalError as e:
-            print(f"Failed to add column '{column_name}' to table '{table_name}': {e}")
-        finally:
-            cursor.close()
+
+    def get_existing_columns(self, table_name):
+        """
+        Get the list of existing columns in a table.
+
+        :param table_name: Name of the table.
+        :return: A set of existing column names.
+        """
+        conn = self.connect()
+        cursor = conn.cursor()
+
+        # Fetch all column names for the table
+        cursor.execute(f"PRAGMA table_info({table_name});")
+        columns_info = cursor.fetchall()
+
+        # Extract column names from the info and close the connection
+        existing_columns = {info[1] for info in columns_info}
+        conn.close()
+        return existing_columns
+
+    def add_column_if_not_exists(self, table_name, column_name, column_type="TEXT"):
+        # Query to check if the column exists
+        conn = self.connect()
+        cursor = conn.cursor()
+        query = cursor.execute(f"PRAGMA table_info({table_name})")
+        columns = [row[1] for row in query.fetchall()]
+
+        # Check if the column is already in the table
+        if column_name not in columns:
+            # If not, add the column
+            cursor.execute(
+                f"ALTER TABLE {table_name} ADD COLUMN {column_name} {column_type};"
+            )
+
+    def check_table_exists(self, table_name):
+        """
+        Checks if a table exists in an SQLite3 database.
+
+        :param db_name: Name of the database file.
+        :param table_name: Name of the table to check.
+        :return: True if the table exists, False otherwise.
+        """
+        # Connect to the SQLite database (or create it if it doesn't exist)
+        conn = self.connect()
+        cursor = conn.cursor()
+
+        # Query to check if the table exists
+        cursor.execute(
+            """
+            SELECT name 
+            FROM sqlite_master 
+            WHERE type='table' AND name=?;
+        """,
+            (table_name,),
+        )
+        result = cursor.fetchone()
+        conn.close()
+        return result is not None
 
 
 # todo -> maybe move elsewhere:
